+++ src/main/java/fr/christophe/cinema/seance/SeanceRepository.java
@@ -2,5 +2,5 @@ package fr.christophe.cinema.seance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeanceRepository extends JpaRepository<Seance, Integer> {
+public interface SeanceRepository extends JpaRepository<Seance, Long> {
 }
